requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.6
django-celery==3.1.17
django-redis==5.4.0
kombu==5.4.2
pillow==11.1.0
prompt_toolkit==3.0.50
PyMuPDF==1.25.3
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.2.1
setuptools==75.8.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
